Normalize lazy service overrides for Windows ESM imports
The lazy plugin service path can be supplied as a drive-letter module override. Reusing the existing import-specifier normalization keeps the default dynamic import aligned with the plugin loader without changing non-Windows specifiers.

Constraint: Node ESM rejects raw Windows drive-letter import specifiers with ERR_UNSUPPORTED_ESM_URL_SCHEME
Rejected: Duplicate the Windows path normalization inside lazy-service-module.ts | duplicating the loader helper would let the two override paths drift
Confidence: high
Scope-risk: narrow
Directive: Keep lazy service override imports and plugin loader imports on the same normalization helper
Tested: pnpm exec oxfmt --check --threads=1 CHANGELOG.md src/plugins/import-specifier.ts src/plugins/lazy-service-module.ts src/plugins/lazy-service-module.test.ts src/plugins/loader.ts
Tested: pnpm test src/plugins/lazy-service-module.test.ts
Tested: pnpm test src/plugins/loader.test.ts -- -t converts Windows absolute import specifiers
Tested: pnpm build
Tested: pnpm check:changed
Not-tested: Manual Windows runtime smoke
