+++ src/plugins/lazy-service-module.ts
@@ -1,4 +1,5 @@
 import { isTruthyEnvValue } from "../infra/env.js";
+import { toSafeImportPath } from "./import-specifier.js";
 
 type LazyServiceModule = Record<string, unknown>;
 
@@ -34,7 +35,8 @@ export async function startLazyPluginServiceModule(params: {
   const overrideEnvVar = params.overrideEnvVar?.trim();
   const override = overrideEnvVar ? process.env[overrideEnvVar]?.trim() : undefined;
   const loadOverrideModule =
-    params.loadOverrideModule ?? (async (specifier: string) => await import(specifier));
+    params.loadOverrideModule ??
+    (async (specifier: string) => await import(toSafeImportPath(specifier)));
   const validatedOverride =
     override && params.validateOverrideSpecifier
       ? params.validateOverrideSpecifier(override)
